reorganizing play.js, created config folder
create config folder to keep from cluttering index.js

moved distube object and event listeners from play.js to config/playercfg.js

moved queue status template from play.js to queue.js

import music player into index.js

client is now being exported from index.js so that client.player (and therefore things like the music queue) can be accessed by commands that use the music player.
