+++ index.js
@@ -1,7 +1,6 @@
 const fs = require('node:fs');
 const path = require('node:path');
 const { Client, Collection, Events, GatewayIntentBits } = require('discord.js');
-
 const { dotenv } = require('dotenv/config');
 
 const client = new Client({ intents: [GatewayIntentBits.Guilds] });
@@ -11,6 +10,13 @@ const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('
 
 const eventsPath = path.join(__dirname, 'events');
 const eventFiles = fs.readdirSync(eventsPath).filter(file => file.endsWith('.js'));
+
+const configPath = path.join(__dirname, 'config');
+
+const musicConfig = require(path.join(configPath, 'playercfg.js'));
+
+// create music player
+client.player = musicConfig;
 
 // read through all the files and add them to a discord.js collection object
 client.commands = new Collection();
@@ -37,3 +43,5 @@ for (const file of eventFiles) {
 }
 
 client.login(process.env.DISCORD_TOKEN);
+
+module.exports = { client };
